Split out PinChangeInt.h into header and cpp source Sometimes your sensor nodes need to respond to interrupts too!
